+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Date format: DD/MM/YYYY
 - Progress Indicators velocity is no longer affected by device frame rate ([#502](https://github.com/bdlukaa/fluent_ui/pull/502)) 
 - Added `AutoSuggestBox.enabled` ([#504](https://github.com/bdlukaa/fluent_ui/issues/504))
 - Correctly keep the `NavigationView` animation state ([cf0fae1](https://github.com/bdlukaa/fluent_ui/commit/cf0fae16ce9a8879653606571e90af12f711bb84) ,[bd89ba6](https://github.com/bdlukaa/fluent_ui/commit/bd89ba6791dfe2762b37b3e291d8d1a7979cb3f5))
+- Calculate `selected` for all parents as soon as the `TreeView` is built
 
 ## [4.0.0-pre.3] - Top navigation and auto suggestions - [13/08/2022]